src/requirements.txt
azure-core
azure-ai-documentintelligence
streamlit
python-dotenv
azure-storage-blob
